feature_extraction/requirements.txt
numpy==1.26.4
Pillow==10.3.0
PyYAML==6.0.1
torch
git+https://github.com/openai/CLIP.git
ftfy
regex
tqdm
loguru